Split the historical registry import into parser collaborators
Pure move gone three deep: a straight parser extraction would have
carried the god-class disease along, so the CSV orchestration, the
per-row field parsing, and the locale-aware value conversion each get
their own class, with ParsedRow promoted to a shared package-private
record. Twenty-two parsing tests leave the Spring Boot integration test
for a plain unit test, and two pre-existing lambda-only call edges are
made direct, which improves cohesion outright.

God classes 6 -> 5, PMD violations 137 -> 135, disconnected classes
454 -> 453.
