+++ metrics/baseline.json
@@ -1,5 +1,5 @@
 {
   "modulithViolations": 0,
   "moduleCycles": 0,
-  "disconnectedClasses": 454
+  "disconnectedClasses": 453
 }
