+++ metrics/scorecard.json
@@ -1,18 +1,18 @@
 {
   "branchCoverage": 87.66,
-  "classCouplingViolations": 35,
+  "classCouplingViolations": 34,
   "clockViolations": 0,
   "cognitiveComplexityViolations": 20,
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
-  "disconnectedClasses": 454,
+  "disconnectedClasses": 453,
   "fqnViolations": 0,
-  "godClasses": 6,
-  "interactionVerifyingTestClasses": 361,
-  "interactionVerifyingTestShare": 36.1,
-  "lineCoverage": 84.03,
-  "longClasses": 3,
+  "godClasses": 5,
+  "interactionVerifyingTestClasses": 362,
+  "interactionVerifyingTestShare": 36.2,
+  "lineCoverage": 84.04,
+  "longClasses": 2,
   "meanInstability": 0.473,
   "mocklessTestClasses": 462,
   "mocklessTestShare": 46.2,
@@ -26,23 +26,23 @@
   "overMockedTestClasses": 111,
   "overMockedTestShare": 11.1,
   "pitestSlicedModules": 52,
-  "pmdViolations": 137,
+  "pmdViolations": 135,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
-    "CouplingBetweenObjects": 35,
+    "CouplingBetweenObjects": 34,
     "CyclomaticComplexity": 72,
-    "GodClass": 6,
+    "GodClass": 5,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 118.0,
+  "slowest10TestClassesSeconds": 115.9,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 165,
-  "testClassesOver10Seconds": 5,
-  "testClassesTotal": 1000,
+  "suiteWallTimeSeconds": 205,
+  "testClassesOver10Seconds": 4,
+  "testClassesTotal": 1001,
   "testIntegrationShare": 6.7,
   "testSliceShare": 13.8,
   "testUnitShare": 79.5,
-  "top15ClassLines": 6848,
+  "top15ClassLines": 6649,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
