+++ src/main/java/ee/tuleva/onboarding/investment/transaction/ingest/RowParseException.java
@@ -1,0 +1,7 @@
+package ee.tuleva.onboarding.investment.transaction.ingest;
+
+class RowParseException extends RuntimeException {
+  RowParseException(String message) {
+    super(message);
+  }
+}
