+++ src/test/groovy/ee/tuleva/onboarding/investment/transaction/ingest/HistoricalRegistryImportServiceIT.java
@@ -3,10 +3,8 @@ package ee.tuleva.onboarding.investment.transaction.ingest;
 import static ee.tuleva.onboarding.fund.TulevaFund.TUK00;
 import static ee.tuleva.onboarding.fund.TulevaFund.TUK75;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.catchThrowableOfType;
 
 import ee.tuleva.onboarding.investment.transaction.BatchStatus;
-import ee.tuleva.onboarding.investment.transaction.HistoricalImportFormatException;
 import ee.tuleva.onboarding.investment.transaction.HistoricalImportResult;
 import ee.tuleva.onboarding.investment.transaction.InstrumentType;
 import ee.tuleva.onboarding.investment.transaction.OrderStatus;
@@ -173,21 +171,6 @@ class HistoricalRegistryImportServiceIT {
   }
 
   @Test
-  void importWithMultipleBlankBrokerTransactionIdsSucceeds() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-2024-050,EE3600109435,IE00BFG1TM61,,BUY,ETF,1000.00,,2025-03-10 09:00:00,SENT,2025-03-12,,,,,,,,
-        GAS-2024-051,EE3600109435,IE00BFG1TM61,,BUY,ETF,2000.00,,2025-03-11 09:00:00,SENT,2025-03-13,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(result.ordersCreated()).isEqualTo(2);
-  }
-
-  @Test
   void importWithAlreadyPersistedDuplicateBrokerTransactionIdReportsRowError() {
     importService.importCsv(
         """
@@ -210,299 +193,6 @@ class HistoricalRegistryImportServiceIT {
             new HistoricalImportResult.RowError(
                 2, "Duplicate brokerTransactionId: brokerTransactionId=BR-3060"));
     assertThat(orderRepository.count()).isEqualTo(ordersBefore);
-  }
-
-  @Test
-  void importParsesSemicolonDelimiterAndEstonianDecimals() {
-    String csv =
-        """
-        order_id;fund_isin;instrument_isin;transaction_id;transaction_type;instrument_type;order_amount;order_quantity;order_timestamp;order_status;expected_settlement_date;actual_settlement_date;execution_timestamp;executed_quantity;unit_price;total_consideration;net_settlement_amount;commission_amount;comment
-        GAS-2024-020;EE3600109435;IE00BFG1TM61;BR-2020;BUY;ETF;25 000,00;250,000000;2025-03-10 09:00:00;EXECUTED;2025-03-12;;2025-03-10 14:30:00;250,000000;100,10;25 025,00;25 030,00;5,00;
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(result.ordersCreated()).isEqualTo(1);
-    TransactionOrder order = findByComment("BR-2020");
-    assertThat(order.getOrderAmount()).isEqualByComparingTo("25000.00");
-    assertThat(order.getOrderQuantity()).isEqualByComparingTo("250.000000");
-    TransactionExecution execution =
-        executionRepository.findByBrokerTransactionId("BR-2020").orElseThrow();
-    assertThat(execution.getUnitPrice()).isEqualByComparingTo("100.10");
-    assertThat(execution.getTotalConsideration()).isEqualByComparingTo("25025.00");
-  }
-
-  @Test
-  void importResolvesSingleSeparatorAsThousandsUnderSemicolonDelimiter() {
-    String csv =
-        """
-        order_id;fund_isin;instrument_isin;transaction_id;transaction_type;instrument_type;order_amount;order_quantity;order_timestamp;order_status;expected_settlement_date;actual_settlement_date;execution_timestamp;executed_quantity;unit_price;total_consideration;net_settlement_amount;commission_amount;comment
-        GAS-2024-070;EE3600109435;IE00BFG1TM61;BR-2070;BUY;ETF;12.345;;2025-03-10 09:00:00;SENT;2025-03-12;;;;;;;;
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    TransactionOrder order = orderRepository.findByInstrumentIsin("IE00BFG1TM61").getFirst();
-    assertThat(order.getOrderAmount()).isEqualByComparingTo("12345");
-  }
-
-  @Test
-  void importResolvesCommaDecimalConventionAcrossThousandsAndDecimalForms() {
-    String csv =
-        """
-        order_id;fund_isin;instrument_isin;transaction_id;transaction_type;instrument_type;order_amount;order_quantity;order_timestamp;order_status;expected_settlement_date;actual_settlement_date;execution_timestamp;executed_quantity;unit_price;total_consideration;net_settlement_amount;commission_amount;comment
-        GAS-2024-071;EE3600109435;IE00BFG1TM61;BR-2071;BUY;ETF;1.234.567;0.80000;2025-03-10 09:00:00;SENT;2025-03-12;;;;150000000,000;100,000;;;
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    TransactionOrder order = orderRepository.findByInstrumentIsin("IE00BFG1TM61").getFirst();
-    assertThat(order.getOrderAmount()).isEqualByComparingTo("1234567");
-    assertThat(order.getOrderQuantity()).isEqualByComparingTo("0.80000");
-    TransactionExecution execution =
-        executionRepository.findByBrokerTransactionId("BR-2071").orElseThrow();
-    assertThat(execution.getUnitPrice()).isEqualByComparingTo("150000000.000");
-    assertThat(execution.getTotalConsideration()).isEqualByComparingTo("100.000");
-  }
-
-  @Test
-  void importResolvesPeriodDecimalConventionAcrossThousandsAndDecimalForms() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-2024-072,EE3600109435,IE00BFG1TM61,BR-2072,BUY,ETF,12.345,"1,234.56",2025-03-10 09:00:00,SENT,2025-03-12,,,,"100,000",1234.56,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    TransactionOrder order = orderRepository.findByInstrumentIsin("IE00BFG1TM61").getFirst();
-    assertThat(order.getOrderAmount()).isEqualByComparingTo("12.345");
-    assertThat(order.getOrderQuantity()).isEqualByComparingTo("1234.56");
-    TransactionExecution execution =
-        executionRepository.findByBrokerTransactionId("BR-2072").orElseThrow();
-    assertThat(execution.getUnitPrice()).isEqualByComparingTo("100000");
-    assertThat(execution.getTotalConsideration()).isEqualByComparingTo("1234.56");
-  }
-
-  @Test
-  void importWithMissingHeadersThrowsFormatException() {
-    String csv =
-        """
-        order_id,fund_isin
-        GAS-2024-030,EE3600109435
-        """;
-
-    HistoricalImportFormatException exception =
-        catchThrowableOfType(
-            HistoricalImportFormatException.class, () -> importService.importCsv(csv));
-
-    assertThat(exception.getMissingHeaders())
-        .containsExactly(
-            "instrument_isin",
-            "order_timestamp",
-            "order_status",
-            "expected_settlement_date",
-            "comment");
-    assertThat(exception.getRequiredHeaders())
-        .containsExactly(
-            "order_id",
-            "fund_isin",
-            "instrument_isin",
-            "order_timestamp",
-            "order_status",
-            "expected_settlement_date",
-            "comment");
-  }
-
-  @Test
-  void executedRowWithBlankOrderTimestampIsRejected() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9001,EE3600109435,IE00BFG1TM61,BR-9001,BUY,ETF,1000.00,10.000000,,EXECUTED,2025-03-12,,2025-03-10 14:30:00,10.000000,100.00,1000.00,995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=order_timestamp, orderId=GAS-9001"));
-    assertThat(result.ordersCreated()).isZero();
-    assertThat(executionRepository.findByBrokerTransactionId("BR-9001")).isEmpty();
-  }
-
-  @Test
-  void executedRowWithNoExecutionDataIsRejected() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9002,EE3600109435,IE00BFG1TM61,BR-9002,BUY,ETF,1000.00,10.000000,2025-03-10 09:00:00,EXECUTED,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Terminal order missing execution data: orderId=GAS-9002"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void executedRowWithBlankUnitPriceIsRejected() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9003,EE3600109435,IE00BFG1TM61,BR-9003,BUY,ETF,1000.00,10.000000,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,10.000000,,1000.00,995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=unit_price, orderId=GAS-9003"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void executedEtfBuyWithBlankExecutedQuantityIsRejected() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9004,EE3600109435,IE00BFG1TM61,BR-9004,BUY,ETF,1000.00,10.000000,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,,100.00,1000.00,995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=executed_quantity, orderId=GAS-9004"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void executedSellWithBlankExecutedQuantityIsRejected() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9005,EE3600109443,IE0009FT4LX4,BR-9005,SELL,FUND,1000.00,10.000000,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,,100.00,1000.00,995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=executed_quantity, orderId=GAS-9005"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void executedFundBuyWithBlankExecutedQuantityButPopulatedTotalConsiderationSucceeds() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9006,EE3600109443,IE0009FT4LX4,BR-9006,BUY,FUND,5000.00,,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,,100.00,5000.00,4995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(result.ordersCreated()).isEqualTo(1);
-    TransactionExecution execution =
-        executionRepository.findByBrokerTransactionId("BR-9006").orElseThrow();
-    assertThat(execution.getExecutedQuantity()).isNull();
-    assertThat(execution.getTotalConsideration()).isEqualByComparingTo("5000.00");
-  }
-
-  @Test
-  void nonTerminalRowsWithBlankTimestampAndEconomicsSucceed() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9007,EE3600109435,IE00BFG1TM61,,BUY,ETF,1000.00,10.000000,,DRAFT,2025-03-12,,,,,,,,
-        GAS-9008,EE3600109435,IE00BFG1TM61,,BUY,ETF,1000.00,10.000000,,SENT,2025-03-12,,,,,,,,
-        GAS-9009,EE3600109435,IE00BFG1TM61,,BUY,ETF,1000.00,10.000000,,CANCELLED,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(result.ordersCreated()).isEqualTo(3);
-  }
-
-  @Test
-  void terminalFundBuyWithoutInstrumentTypeColumnResolvesTypeFromInstrumentReference() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9101,EE3600109443,IE0009FT4LX4,BR-9101,BUY,5000.00,,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,,100.00,5000.00,4995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(result.ordersCreated()).isEqualTo(1);
-    TransactionOrder order = orderRepository.findByInstrumentIsin("IE0009FT4LX4").getFirst();
-    assertThat(order.getInstrumentType()).isEqualTo(InstrumentType.FUND);
-  }
-
-  @Test
-  void etfBuyWithoutInstrumentTypeColumnStillRequiresExecutedQuantity() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9102,EE3600109443,IE000I9HGDZ3,BR-9102,BUY,5000.00,,2025-03-10 09:00:00,EXECUTED,2025-03-12,,2025-03-10 14:30:00,,100.00,5000.00,4995.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=executed_quantity, orderId=GAS-9102"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void unknownInstrumentIsinWithoutInstrumentTypeColumnIsRejectedWithoutEtfGuess() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9103,EE3600109443,XX0000000099,,BUY,5000.00,,2025-03-10 09:00:00,SENT,2025-03-12,,,,,,,,
-        """;
-    long ordersBefore = orderRepository.count();
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Unknown instrument: instrumentIsin=XX0000000099"));
-    assertThat(result.ordersCreated()).isZero();
-    assertThat(orderRepository.count()).isEqualTo(ordersBefore);
-  }
-
-  @Test
-  void explicitInstrumentTypeColumnWinsOverInstrumentReference() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9104,EE3600109443,IE000I9HGDZ3,,BUY,FUND,5000.00,,2025-03-10 09:00:00,SENT,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    TransactionOrder order = orderRepository.findByInstrumentIsin("IE000I9HGDZ3").getFirst();
-    assertThat(order.getInstrumentType()).isEqualTo(InstrumentType.FUND);
   }
 
   @Test
@@ -561,172 +251,6 @@ class HistoricalRegistryImportServiceIT {
     TransactionExecution execution =
         executionRepository.findByBrokerTransactionId("BR-9203").orElseThrow();
     assertThat(execution.getReportedDate()).isNull();
-  }
-
-  @Test
-  void importReadsTimestampsInIsoInstantIsoLocalAndDateOnlyForms() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9204,EE3600109435,IE00BFG1TM61,BR-9204,BUY,ETF,25000.00,250.000000,2025-03-10T09:00:00Z,EXECUTED,2025-03-12,,2025-03-10T14:30:00Z,250.000000,100.10,25025.00,25030.00,5.00,
-        GAS-9205,EE3600109435,IE00BFG1TM61,BR-9205,BUY,ETF,25000.00,250.000000,2025-03-11T09:00:00,EXECUTED,2025-03-13,,2025-03-11T14:30:00,250.000000,100.10,25025.00,25030.00,5.00,
-        GAS-9206,EE3600109435,IE00BFG1TM61,BR-9206,BUY,ETF,25000.00,250.000000,2025-03-12,EXECUTED,2025-03-14,,2025-03-12,250.000000,100.10,25025.00,25030.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    assertThat(
-            executionRepository
-                .findByBrokerTransactionId("BR-9204")
-                .orElseThrow()
-                .getExecutionTimestamp())
-        .isEqualTo(Instant.parse("2025-03-10T14:30:00Z"));
-    assertThat(
-            executionRepository
-                .findByBrokerTransactionId("BR-9205")
-                .orElseThrow()
-                .getExecutionTimestamp())
-        .isEqualTo(Instant.parse("2025-03-11T14:30:00Z"));
-    assertThat(
-            executionRepository
-                .findByBrokerTransactionId("BR-9206")
-                .orElseThrow()
-                .getExecutionTimestamp())
-        .isEqualTo(Instant.parse("2025-03-12T00:00:00Z"));
-  }
-
-  @Test
-  void importReadsEstonianDates() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9207,EE3600109435,IE00BFG1TM61,BR-9207,BUY,ETF,25000.00,250.000000,2025-03-10 09:00:00,EXECUTED,12.03.2025,,2025-03-10 14:30:00,250.000000,100.10,25025.00,25030.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors()).isEmpty();
-    TransactionOrder order = findByComment("BR-9207");
-    assertThat(order.getExpectedSettlementDate()).isEqualTo(LocalDate.of(2025, 3, 12));
-  }
-
-  @Test
-  void importRejectsAnUnreadableDate() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9208,EE3600109435,IE00BFG1TM61,,BUY,ETF,25000.00,250.000000,,SENT,2025/03/12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Invalid date: column=expected_settlement_date, value=2025/03/12"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsAnUnreadableNumber() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9209,EE3600109435,IE00BFG1TM61,,BUY,ETF,not a number,250.000000,,SENT,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Invalid number: column=order_amount, value=not a number"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsAnUnknownOrderStatus() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9210,EE3600109435,IE00BFG1TM61,,BUY,ETF,25000.00,250.000000,,PARTIALLY_FILLED,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Invalid value: column=order_status, value=PARTIALLY_FILLED"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsARowWithNoOrderId() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        ,EE3600109435,IE00BFG1TM61,,BUY,ETF,25000.00,250.000000,,SENT,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(new HistoricalImportResult.RowError(2, "Missing value: column=order_id"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsASettledRowWithNoSettlementDateOnEitherSide() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9211,EE3600109435,IE00BFG1TM61,BR-9211,BUY,ETF,25000.00,250.000000,2025-03-10 09:00:00,SETTLED,,,2025-03-10 14:30:00,250.000000,100.10,25025.00,25030.00,5.00,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2,
-                "Settled row missing both actual and expected settlement date: orderId=GAS-9211"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsAFundSubscriptionWithNoTotalConsideration() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9212,EE3600109443,LU0826455353,BR-9212,BUY,FUND,5000.00,,2025-05-05 09:00:00,EXECUTED,2025-05-09,,2025-05-05 14:30:00,,100.00,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(
-                2, "Missing value: column=total_consideration, orderId=GAS-9212"));
-    assertThat(result.ordersCreated()).isZero();
-  }
-
-  @Test
-  void importRejectsTheFileWhenTheSameOrderIdAppearsTwice() {
-    String csv =
-        """
-        order_id,fund_isin,instrument_isin,transaction_id,transaction_type,instrument_type,order_amount,order_quantity,order_timestamp,order_status,expected_settlement_date,actual_settlement_date,execution_timestamp,executed_quantity,unit_price,total_consideration,net_settlement_amount,commission_amount,comment
-        GAS-9213,EE3600109435,IE00BFG1TM61,,BUY,ETF,25000.00,250.000000,,SENT,2025-03-12,,,,,,,,
-        GAS-9213,EE3600109435,IE00BFG1TM61,,BUY,ETF,25000.00,250.000000,,SENT,2025-03-12,,,,,,,,
-        """;
-
-    HistoricalImportResult result = importService.importCsv(csv);
-
-    assertThat(result.errors())
-        .containsExactly(
-            new HistoricalImportResult.RowError(3, "Duplicate order_id in file: orderId=GAS-9213"));
-    assertThat(result.ordersCreated()).isZero();
   }
 
   private TransactionOrder findByComment(String brokerTransactionId) {
